+++ api/src/main/java/net/jqwik/api/arbitraries/IntegerArbitrary.java
@@ -28,4 +28,10 @@ public interface IntegerArbitrary extends Arbitrary<Integer> {
 	 * Set the allowed upper {@code max} (included) bounder of generated numbers.
 	 */
 	IntegerArbitrary lessOrEqual(int max);
+
+	/**
+	 * Set shrinking target to {@code target} which must be between the allowed bounds.
+	 */
+	@API(status = EXPERIMENTAL, since = "1.1.4")
+	Arbitrary<Integer> shrinkTowards(int target);
 }